add exception handler for Account

+++ userMS/src/main/java/tpe/userMS/service/AccountService.java
@@ -1,16 +1,18 @@
 package tpe.userMS.service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import jakarta.validation.Valid;
+import jakarta.ws.rs.InternalServerErrorException;
 import tpe.userMS.DTO.DTOAccountRequest;
 import tpe.userMS.DTO.DTOAccountResponse;
 import tpe.userMS.DTO.DTOReduceBalanceRequest;
+import tpe.userMS.exception.AccountWithoutMoneyException;
+import tpe.userMS.exception.NotFoundException;
 import tpe.userMS.model.Account;
 import tpe.userMS.model.User;
 import tpe.userMS.repository.AccountRepository;
@@ -25,46 +27,38 @@ public class AccountService {
 	private UserRepository userRepository;
 	
 	@Transactional(readOnly = true)
-	public DTOAccountResponse getAccountByUserIdWithBalance(long userId) throws Exception, NoSuchElementException {
-		try {
-			return repository.getAccountByUserIdWithBalance(userId)
-					.map( DTOAccountResponse::new )
-					.orElseThrow(() -> new Exception("Accounts associated to user with id" + userId + " don't exist or they do not have balance"));
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
-		}
+	public DTOAccountResponse getAccountByUserIdWithBalance(long userId) throws AccountWithoutMoneyException {
+		return repository.getAccountByUserIdWithBalance(userId)
+				.map( DTOAccountResponse::new )
+				.orElseThrow(() -> new AccountWithoutMoneyException(userId));
 	}
 
 	@Transactional(readOnly = true)
-	public List<DTOAccountResponse> findAll() throws Exception {
+	public List<DTOAccountResponse> findAll() throws InternalServerErrorException {
 		try {
 			return repository.findAll().stream().map(DTOAccountResponse::new).toList();
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
+		} catch (InternalServerErrorException e) {
+			throw new InternalServerErrorException(e.getMessage());
 		}
 	}
 
 	@Transactional
-	public DTOAccountResponse save(@Valid DTOAccountRequest request) throws Exception {
+	public DTOAccountResponse save(@Valid DTOAccountRequest request) {
 		try {
 			Account account = new Account(request.getId(), request.getMoneyBalance());
 			account = repository.save(account);
 			return new DTOAccountResponse(account);
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
+		} catch (InternalServerErrorException e) {
+			throw new InternalServerErrorException(e.getMessage());
 		}
 	}
 
 	
 		@Transactional(readOnly = true)
-		public DTOAccountResponse getAccountById(long id) throws Exception {
-		    try {
-		        return repository.getAccountById(id)
-		                .map(DTOAccountResponse::new)
-		                .orElseThrow(() -> new Exception("Account not found for id: " + id));
-		    } catch (Exception e) {
-		        throw new Exception(e.getMessage());
-		    }
+		public DTOAccountResponse getAccountById(long id) throws NotFoundException {
+	        return repository.getAccountById(id)
+	                .map(DTOAccountResponse::new)
+	                .orElseThrow(() -> new NotFoundException("Account", id));
 		}
 
 		
@@ -78,48 +72,60 @@ public class AccountService {
 	}
 
 	@Transactional
-	public void delete(long id) throws Exception {
+	public void delete(long id) throws NotFoundException  {
 		try {
-			repository.deleteAccountById(id);
+			if (repository.existsById(id)) {
+				repository.deleteAccountById(id);
+			} else {
+				throw new NotFoundException("Account", id);
+			}
 		} catch (NumberFormatException e) {
-	        throw new Exception("Invalid ID format");
+	        throw new NumberFormatException("Invalid ID format");
 		}
 	}
 	
 	@Transactional
-    public void reduceMoneyBalance(long id, DTOReduceBalanceRequest request) throws Exception {
-        try {
-            repository.reduceAccountMoneyBalance(request.getMoney(), id);
-        } catch (Exception e) {
-            throw new Exception("Failed to reduce account money balance: " + e.getMessage());
-        }
+    public void reduceMoneyBalance(long id, DTOReduceBalanceRequest request) throws NotFoundException {
+		try {
+			if (repository.existsById(id)) {
+				repository.reduceAccountMoneyBalance(request.getMoney(), id);
+			} else {
+				throw new NotFoundException("Account", id);
+			}
+		} catch (InternalServerErrorException e) {
+			throw new InternalServerErrorException(e.getMessage());
+		}
     }
 
 
     @Transactional
-    public void updateMoneyBalance(long id, int money) throws Exception {
+    public void updateMoneyBalance(long id, int money) throws NotFoundException  {
         try {
-            repository.updateAccount(money, id);
-        } catch (Exception e) {
-            throw new Exception("Failed to update account money balance: " + e.getMessage());
+        	if (repository.existsById(id)) {
+        		repository.updateAccount(money, id);
+        	} else {
+        		throw new NotFoundException("Account", id);
+        	}
+        } catch (InternalServerErrorException e) {
+            throw new InternalServerErrorException("Failed to update account money balance: " + e.getMessage());
         }
     }
 
 	@Transactional
-	public void addUserToAccount(Long id, Long userId) throws Exception {
+	public void addUserToAccount(Long id, Long userId) throws NotFoundException {
 		try {
 			User user = userRepository.findById(userId)
-					.orElseThrow(() -> new RuntimeException("User not found with userId: " + userId));
+					.orElseThrow(() -> new NotFoundException("User", userId));
 			Account account = repository.findById(id)
-					.orElseThrow(() -> new RuntimeException("Account not found with id: " + id));
+					.orElseThrow(() -> new NotFoundException("Account", id));
 //			account.getUsers().add(user);
 //			repository.save(account); No funciona porque user es el duenio de la relacion
 			
 			user.getAccounts().add(account);
 			userRepository.save(user);
 			
-		} catch (Exception e) {
-			throw new Exception("Failed to update account user: " + e.getMessage());
+		} catch (InternalServerErrorException e) {
+			throw new InternalServerErrorException("Failed to update account user: " + e.getMessage());
 		}
 	}
